reorganize the defintions, adds definitions for addons including OCRP and PDF
